requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Google Cloud SDK
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
google-cloud-bigquery>=3.11.0
google-cloud-pubsub>=2.18.0
google-cloud-scheduler>=2.10.0
google-cloud-run>=0.10.0
google-cloud-functions>=1.13.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.95.0

# Vertex AI and ML
google-cloud-aiplatform>=1.35.0
vertexai>=1.46.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
aiofiles>=23.2.0

# Task Queue and Async
celery>=5.3.0
redis>=5.0.0
asyncio>=3.4.3
aioredis>=2.0.1

# Data Processing
pandas>=2.1.0
numpy>=1.25.0
pyarrow>=14.0.0

# Monitoring and Logging
prometheus-client>=0.18.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development Tools
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# CLI and Utils
click>=8.1.0
rich>=13.6.0
tabulate>=0.9.0
humanize>=4.8.0